Change configuration parameter from double to float for FundamentalMatrixEstimation

+++ interfaces/SolARFundamentalMatrixEstimationOpencv.h
@@ -55,14 +55,14 @@ private:
     bool isFValid(const Transform2Df & F);
 
     /// @brief The desirable level of confidence (propability) that the estimated matrix is correct.
-    double m_confidenceLevel = 0.99;
+    float m_confidenceLevel = 0.99;
 
     ///  @brief threshold to define which point are ouliers
     ///  Here we are using a RANSAC method to remove outlier.
     ///  This attribute is the ratio between the maximum distance in pixels between source points and the maximum distance in pixels to the epipolar line for which point is considered as a outlier.
     ///  The higher is this ratio, the more you will keep inliers to estimate your 2D transform, but the less this estimation will be correct.
     ///  By default, this value is set to the one proposed by [Snavely07 4.1]
-    double m_outlierDistanceRatio = 0.006;
+    float m_outlierDistanceRatio = 0.006;
 };
 
 }
